# docs/openapi.yml
openapi: 3.1.0
info:
  title: Tracecat API
  summary: Tracecat API
  description: Tracecat is the open source Tines / Splunk SOAR alternative. You can
    operate Tracecat in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Tracecat Founders
    email: founders@tracecat.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /settings/app:
    get:
      tags:
      - settings
      summary: Get App Settings
      operationId: settings-get_app_settings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettingsRead'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
    patch:
      tags:
      - settings
      summary: Update App Settings
      operationId: settings-update_app_settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppSettingsUpdate'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
...
  /auth/oauth/callback:
    get:
      tags:
      - auth
      summary: Oauth:Google.Database.Callback
      description: The response varies based on the authentication backend used.
      operationId: auth-oauth:google.database.callback
      parameters:
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: code_verifier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code Verifier
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State
      - name: error
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          content:
            application/json:
              examples:
                INVALID_STATE_TOKEN:
                  summary: Invalid state token.
                LOGIN_BAD_CREDENTIALS:
                  summary: User is inactive.
                  value:
                    detail: LOGIN_BAD_CREDENTIALS
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad Request
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AppSettingsRead:
      properties:
        app_registry_validation_enabled:
          type: boolean
          title: App Registry Validation Enabled
        app_executions_query_limit:
          type: integer
          title: App Executions Query Limit
        app_interactions_enabled:
          type: boolean
          title: App Interactions Enabled
      type: object
      required:
      - app_registry_validation_enabled
      - app_executions_query_limit
      - app_interactions_enabled
      title: AppSettingsRead
      description: Settings for the app.
    AppSettingsUpdate:
      properties:
        app_registry_validation_enabled:
          type: boolean
          title: App Registry Validation Enabled
          description: Whether registry validation is enabled.
          default: false
        app_executions_query_limit:
          type: integer
          title: App Executions Query Limit
          description: The maximum number of executions to return in a single query.
          default: 100
        app_interactions_enabled:
          type: boolean
          title: App Interactions Enabled
          description: Whether app interactions are enabled.
          default: false
      type: object
      title: AppSettingsUpdate
      description: Settings for OAuth authentication.
...
    ErrorModel:
      properties:
        detail:
          anyOf:
          - type: string
          - additionalProperties:
              type: string
            type: object
          title: Detail
      type: object
      required:
      - detail
      title: ErrorModel
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-tracecat-service-key
...